Adding notebook to convert hadiths into Q/A pairs
To fine-tune chat-based models, examples need to be provided in the form of a (prompt, response pair). So the original list of hadiths need to be mapped to a prompt/response pair. An LLM is used to take a hadith as input and generate a prompt such that the hadith is an answer to that question.